requirements.txt
# --- Core API Server ---
fastapi==0.115.5
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# --- Database ---
sqlalchemy==1.4.53
alembic==1.13.1
psycopg2-binary==2.9.9

# --- Data Validation ---
pydantic==2.10.3
pydantic-settings==2.6.1

# --- AI & Utilities ---
google-generativeai>=0.8.3
httpx==0.28.1
slowapi==0.1.9
cachetools==5.5.0
deep-translator==1.11.4
python-dotenv==1.0.0

# --- Security ---
cryptography==44.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4